feat: Add MongoDB connection and authentication to login route
This commit adds the necessary code to establish a connection to MongoDB and authenticate users in the login route. The `pymongo` library is imported and used to connect to the MongoDB Atlas cluster. The connection string is stored in the `MONGO_URI` variable. If the connection is successful, the `users_collection` is accessed to find a user with the submitted credentials. If a matching user is found, the user is authenticated and redirected to the dashboard. Otherwise, an error message is displayed on the login page.